Resume_critique/pyproject.toml
[project]
name = "resume-critique"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.79.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
]
